Upgrade pom, usage types, add tests, UI theme
Several coordinated backend and frontend updates:

- pom.xml: set Spring Boot parent to 3.2.5, target Java 17, add lombok.version, adjust lombok scope/version and replace webmvc/test deps (added spring-security-test).
- Backend fixes: replace Locale.of(...) with new Locale(...); remove stray comment in DearFutureApplication.
- Usage API: change UsageResponse fields from int to long and update UsageServiceImpl to use long for counting/remaining computations.
- Tests: add multiple controller/service unit test skeletons under src/test (Auth, ContactUs, Message, Subscription, User, UsageServiceImpl and TestSecuritySupport).
- Frontend: switch welcome page to dark theme and improve accessibility; add topbar header in App.jsx; introduce feature list and icon usage in WelcomePage.jsx.
- Styling refactor: big CSS overhaul to use theme variables (theme.css additions), unify spacing/typography tokens, and modernize Sidebar/Footer/index/WelcomePage styles.

These changes align typesafety for usage counters, update platform dependencies, add tests, and centralize UI theming for consistent styling.

+++ DearFuture/src/main/java/example/DearFuture/user/dto/response/UsageResponse.java
@@ -19,11 +19,11 @@ public class UsageResponse {
     private String planCode;
     private String planName;
     /** Bu dönem veya toplam maksimum mesaj hakkı */
-    private int limit;
+    private long limit;
     /** Kullanılan mesaj sayısı */
-    private int used;
+    private long used;
     /** Kalan hak */
-    private int remaining;
+    private long remaining;
     /** Her ay yenileniyor mu (FREE: false, PLUS/PREMIUM: true) */
     private boolean resetsMonthly;
     /** Dönem başlangıcı (FREE için null) */
